requirements.txt
metabase-api
baserowapi
python-dotenv